chore(lint): clippy-clean workspace + shared [workspace.lints] baseline
Fixes all 47 clippy findings (auto-fix + hand fixes): Reverse-key sorts,
checked divisions, boxed ureq error (result_large_err), Option-ized audioviz
connect, PtySession type alias, ArchiveStore::is_empty, let-else → `?`,
doc-list spacing. Deliberate escapes are per-site #[allow] with a reason
(boost_save's IPC arg list, cfg-gated Chrome roots, the -sys FFI module,
CLI stdout in main).

Adds [workspace.lints] (inherited by every crate via [lints] workspace=true):
clippy defaults + dbg_macro / todo / print_stdout / unsafe_op_in_unsafe_fn,
so `cargo clippy --workspace` staying at zero is the bar.

Also registers the fmt baseline commit in .git-blame-ignore-revs
(`git config blame.ignoreRevsFile .git-blame-ignore-revs` to activate).

Full workspace suite green; zero clippy warnings.

+++ crates/flux-term/src/grid.rs
@@ -367,8 +367,8 @@ impl vte::Perform for Grid {
             // CUP — cursor position (1-based in the protocol).
             'H' => {
                 let mut it = params.iter();
-                let row = it.next().map_or(1, |p| p[0].max(1)) as u16;
-                let col = it.next().map_or(1, |p| p[0].max(1)) as u16;
+                let row = it.next().map_or(1, |p| p[0].max(1));
+                let col = it.next().map_or(1, |p| p[0].max(1));
                 self.cursor = ((col - 1).min(self.cols - 1), (row - 1).min(self.rows - 1));
             }
             // ED — erase display (mode 2: all).
